Ardiuno stress test.
Create a Notehub project - build this binary and copy to 3 different
filenames, and upload as 3 different host DFU binaries.

Setup thie project with the environment variable host_binary_list
containing a list of the 3 binary names as assigned by notehub:

host_binary_list sketch_hw_dfu_1$20220504144612.bin,sketch_hw_dfu_2$20220504144622.bin,sketch_hw_dfu_3$20220504144631.bin

Configure the Notecard with the host and project UIDs and set to
periodic or continuous mode then let the test run.

To monitor the process, and for failure evaluation on the notecard USB type

trace +req
trace -sync

Then let the test run for hours or days.

In theory this test should work with an ESP32 host, but it has only been
tested with a Swan.